edit all text content and move file directories
