+++ webpack.config.js
@@ -15,10 +15,6 @@ module.exports = {
         use: ['style-loader', 'css-loader', 'sass-loader'],
       },
       {
-        test: /\.(woff|woff2|eot|ttf|otf)$/i,
-        type: 'asset/resource',
-      },
-      {
         test: /\.(jpg|png|gif|svg)$/i,
         use: [
           {
